+++ lib/Utils/MultiSelectorListStyle.dart
@@ -1,1 +1,0 @@
-// enum MultiSelectorListStyle { LIST, CHIP }