+++ lib/flutter_multi_selector.dart
@@ -1,5 +1,4 @@
 export 'Utils/MultiSelectorActions.dart';
 export 'Utils/MultiSelectorItem.dart';
-export 'Utils/MultiSelectorListStyle.dart';
 export 'DialogBox/MultiSelectorDialog.dart';
 export 'DialogBox/MultiSelectorDialogField.dart';