+++ test/flutter_multi_selector_test.dart
@@ -1,8 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:flutter_multi_selector/flutter_multi_selector.dart';
-
-void main() {
-
-
-}